engine/ui: text input widget + host text-input bridge
Adds the TextInput widget (caret blink, horizontal scroll, focus,
keyboard editing) and the supporting infrastructure:

- engine/ui: ImGui-style vertical layout pen (LayoutNextRow, Avail);
  WindowAutoSizeY patches bg+clip from measured pen at EndWindow;
  click-outside focus clear; KeyEvent/Text on InputFrame.
- engine/backend: TextInput{Begin,End,Poll} shims — WASM hidden <input>
  overlay drives IME/clipboard, native sokol stubs (SDL3 swap-ready).
- engine/text: per-widget volatile line raster path (DrawVolatile/
  DropVolatile) so focused inputs reshape in place without churning
  the atlas; text.Font.Metrics exposes ascent/descent/lineheight.
- backend: bubble html5 key/char events so canvas keeps receiving keys
  while the hidden input has focus.
- game: ui overlay stacks LogView + two TextInputs in one auto-sized
  HUD window.
- ai/ui-design.md, ai/text-input.md, ai/text-review.md, CLAUDE.md
  updated to reflect the new widget + layout model.

Made-with: Cursor

+++ engine/text/atlas_native.go
@@ -141,7 +141,10 @@ func (a *glyphAtlas) shapeRun(h int32, s string, ascentPx int32) ([]lineQuad, em
 	}
 	var out []lineQuad
 	penX := int32(0)
-	atlasF := a.atlasSize
+	// Convert to float ONCE; UV math must run in float space — the original
+	// `entry.x / atlasF` was int32 division and collapsed to 0 for every
+	// glyph (entry.x < atlasSize), making every quad sample atlas pixel (0,0).
+	atlasF := float32(a.atlasSize)
 	for _, glyph := range shaped {
 		entry, err := a.ensureGlyph(h, glyph.GlyphID)
 		if err != nil {
@@ -159,10 +162,10 @@ func (a *glyphAtlas) shapeRun(h int32, s string, ascentPx int32) ([]lineQuad, em
 		out = append(out, lineQuad{
 			Dst: emath.Rect{X: x0, Y: y0, W: w, H: hgt},
 			UV: emath.UVRect{
-				U0: float32(entry.x / atlasF),
-				V0: float32(entry.y / atlasF),
-				U1: float32(entry.x + entry.bitmap.WidthPx/atlasF),
-				V1: float32(entry.y + entry.bitmap.HeightPx/atlasF),
+				U0: float32(entry.x) / atlasF,
+				V0: float32(entry.y) / atlasF,
+				U1: float32(entry.x+entry.bitmap.WidthPx) / atlasF,
+				V1: float32(entry.y+entry.bitmap.HeightPx) / atlasF,
 			},
 		})
 		penX += glyph.XAdvance26_6
